Added ReadContext class to carry state around
This is in preparation for 377. Wanted to submit this first before doing 377, as this is pure refactoring with no functional change. Intent is to simplify all the classes that need state passed to them - we toss everything into ReadContext (and ensure it's Serializable), which means we can add functionality in there too.