DEVEXP-378 Now using multiple hosts in a cluster
I wrote up DEVEXP-406 to address what happens when a particular host stops working; DMSDK has support for that - i.e. that client is removed and not used again. 

To fully test this, we need a multi-node cluster, and ideally with at least 1 node not having a forest for the database. That will be addressed in a follow-up PR that adds a docker-compose file for building such a cluster and doing manual testing with it.