DEVEXP-430 Fix for streaming test
Tweaked logging a bit too, and also removed Serializable from MarkLogicWrite - I had added that in my prototype, but I think I did that based on something wrong I was doing. It does not appear needed, which makes sense - the point of the Write object is to produce a serializable Factory that can be sent to a different Spark host.