DEVEXP-459 Now using the latest server timestamp on each separate query
The primary change is that `ReadContext` is now created in the scan builder instead of in `DefaultSource`. This ensures a new one is created each time the user queries for data. 

Added some debug logging so it's easy to see when certain 
objects are created. 

Also switched from `resultRows` to `resultDoc`. Going to create a bug for the Java Client as `resultRows` ignores server timestamp as an input.