Fixed statement in query restrictions
The restrictions on RowBatcher for e.g. groupBy/limit/orderBy don't apply to the Spark connector. The connector will push those down to MarkLogic, but if more than 1 call is made to MarkLogic, Spark will apply them as well, thus avoiding the problem that RowBatcher has where those calls only get applied to each batch and not the entire set of matching rows.