Bumped Spark to 3.4.1
This reverts the change from 1 citations doc to 15 author docs, as that was producing some odd test failures on MarkLogic 11 that need further investigation. I had bumped to Spark 3.5.0, but that yielded further errors. All tests are passing on 3.4.1, so sticking with that for now, with a goal of getting to 3.5.0 for the 2.1.0 release. 

Also made the numbers more realistic in AnalyzePlanTest.

And removed the volumes from docker-compose.yaml so it's easier to start from scratch.