Tweaked logging in MarkLogicWrite
Via some testing, I realized that with large numbers of partitions, logging all the commit messages isn't a great idea. So bumped that to debug, and made the abort messages at the "warn" level (as if the job was aborted, it was either intentional, or an error occurred which will already have copius logging).