Converted test documents to be 15 separate author documents
This is to allow for the tests to pass on MarkLogic 12. Functionally, the tests pass - except that the assertions on how many rows were returned from MarkLogic (which can differ from the number of rows returned to the user) would fail on MarkLogic 12 because all rows are in the same partition, since they're in the same document.