Fixed issue with Jackson dependencies
Ran into an issue when testing the java-dependency project, where the Java Client's preferred Jackson version of 2.15.2 was being used, which Spark 3.4 does not like. Made me realize that we don't want our connector to bring any Jackson dependencies with it - it needs to use whatever the user's Spark distribution is using. So updated our build.gradle file to exclude the Jackson dependencies coming from java-client-api, ml-app-deployer, and marklogic-unit-test (the latter two ensure that we don't get any Jackson 2.15.2 dependencies when running the tests, but get Spark 3.4.1's preferred version instead, which is 2.14.2). 

After fixing this, I realized it's perfectly safe to use Java Client 6.4.0, as that was using Jackson 2.15.2 just like Java Client 6.3.0. So made that upgrade. All tests are passing, and verified that all examples in the docs work as well.