Added note about Spark version support
And modified pushdown tests to not verify the number of rows read from MarkLogic when testing with Spark 3.5.0. The apparent change in Spark's pushdown support will be addressed in a future release. For now, this change to the tests provides an easy way to verify that the tests run correctly against Spark 3.3, 3.4, and 3.5.