Added tests to verify that serializable objects are serializable
Turns out that they are serializable, but I wasn't certain as I thought no-arg constructors were required - but that doesn't appear to be the case based on what `SerializedUtil` does, and Spark does the same thing for its default serialization strategy. 

I made one impl change, though it has no functional impact - I modified ReadContext so that it's not trying to access an active SparkSession, which was making testing more difficult.