DEVEXP-673 Can now read zip files
Main thing of note here is that we extend the Spark `FileTable` class, which is extended by all of Spark's OOTB file-based data sources. This is the only way I could find to reuse the super-valuable code for creating a file index, which applies all of Spark's support for finding files based on one or more paths. 

And since `MarkLogicFileTable` is a `Table`, this becomes a new batch of reader components that branches off from `DefaultSource`. We will likely do the same thing when we support "Read documents from MarkLogic" instead of trying to shoehorn that into the current Optic-based reader.