Can now write documents as files
Moved `MarkLogicFileTable` to base package so it can be reused for both reading and writing. 

`DocumentFileWriter` is the main code; the rest is Spark plumbing. It borrows chunks of code from MLCP for writing a file. 

We'll be updating the Factory here with additional writers very soon - e.g. a compressed file writer, an archive file writer, etc.