Couple fixes pertaining to batch size
Made a nicer error for when batch size is invalid.

More importantly, changed how documents are retrieved to be done via a POST instead of a GET, thereby avoiding Request-URI length errors. Unfortunately was not able to create a test for this - was not able to generate the error locally.