MLE-12523 Ignoring most file errors
This is already pretty big; going to do a follow up PR for aggregate XML zips, which is complex.

The general approach here is:

1. Instead of getting the next row in `next()` in a reader, get it in `hasNext()`. This allows us to catch any error and log it and return false instead of throwing it (we're not allowed to return null from `next()`).
2. `next()` then just returns the next object that was calculated during `hasNext()`. 

Aggregate XML files are particularly complicated because we need to handle child-level errors - such as a URI element missing - where we log the error and keep processing other child elements. That isn't finished in this PR, will finish it in follow up PR. 

Couple simplifications in here too:

1. `AbstractRdfStreamReader` handles everything except whether to open an iterator on triples or quads; the subclasses handle that.
2. `fileContext.open` now handles gzipped files.