MLE-12822 Added tests for Spark JSON
This shows how Spark JSON can be used for the common use case of reading each object in an array in a JSON file as a separate row, and then writing each row to MarkLogic as a separate document. So we don't need our own file reader for this. 

I retained some naming improvements in `AggregateXmlSplitter`, but those are non-functional changes.