MLE-12850 Refactor of code for builing document rows
Started to duplicate this again for 12850 and decided to refactor via the new `DocumentRowBuilder` class. 

Also fixed one bug - the "format" column should not be XML when reading from an archive zip; we don't know what the format is because there's no indication in the zip. 

It is possible to populate "format" for MLCP archives because the MLCP metadata does contain a "format" element.