MLE-13771 Fixing bug where sqlCondition was added to op.and
Was able to reproduce the bug by including 2 filters, one of which produced a `sqlCondition`. I was not able to reproduce an issue where an `and` is created with a `sqlCondition` on the left or right. Have not figured out a way to force Spark to pass that to our connector.