Fixed bug with empty archives being created
The issue was that in order to "fail fast" if the user-defined archive path for failed document is invalid, we had to try to create a zip file first. But that results in an empty zip file if no errors occur. Deleting that is problematic, as the user may not have permission to do so, particularly for an S3 path. 

So we no longer fail fast on an invalid user-defined archive path, in favor of avoiding creating empty archive files.