MLE-13855 Can now capture complex document properties in a document row
This changes the schema of a document row, but that schema was faulty in that complex document properties could not be handled in a map expecting string keys and string values. So the API of a document row changes with this, but it does so to fix a bug and allow for any document properties content to be handled correctly.