threadCount now refers to total thread count
Will communicate this change in the release notes, as it's changing how a public option works, but doing so in a way that better matches user expectations. `threadCountPerPartition` is then offered for what is expected to be a rare case where a user wants to configure the number of threads per partition.