Fixed progress logging bug
Realized in a Spark environment, the static for counting written docs wasn't getting reset. Did some refactoring to avoid duplicated code as well, and also to setup for potentially logging progress on reads.

Also defaulting the progress interval to 0 instead of 10000. This doesn't seem as useful in a Spark environment. Flux will default it to 10000 instead.