Now logging progress for read operations
Had to make two separate loggers so we can have a static counter for reading and a static counter for writing. Both get reset when DefaultSource is created. 

Did some quick manual testing of this in Flux, and it "just works". 

Note that progress for reading/writing files is not being addressed yet. Just focusing on progress for reading from MarkLogic and writing to MarkLogic.