Updated Sonar to latest version
My local Sonar on 10.2-1 crapped out locally. Couldn't revive it. But then found that the latest sonarqube - 10.6 - is running just fine, though it requires Java 17.

Bumped up some test dependencies and removed the "signing" plugin which is no longer needed.

Also removed the dependency check plug-in, as we're not making use of it yet.