MLE-17232 Ensuring logging works in Flux
slf4j-api 1.7.x was sneaking in from the Java Client. The Java Client will be updated to use 2.x. But this ensures that all dependencies use slf4j-api 2.x, which avoids the logging issue referenced in the comment.

Also made a small optimization in shrinking the shadow jar size - we don't need any of the Jackson core dependencies for dataformat-xml, as Spark provides all of them.