MLE-17216 Can now split XML documents
This gets the ball rolling by allowing a user to select text in an XML document and add chunks to the same document. Future stories will support JSON, different splitters, and different storage strategies. 

Going to follow this up with a small PR for Flux so that we can immediately have this available in a Flux snapshot build.