Refactor: Extracted marklogic-spark-api
We now have a loose coupling between the connector and the langchain4j stuff. For the 2.5.0 release, it'll all be combined, but this prepares us for upgrading langchain4j in the near future.

Haven't gotten jacoco working across multiple subprojects. That's an issue in Flux too. Going to figure that out soon.