Tweaked calculation of buckets per partition
While researching the effects of batch size = zero, I realized that it would be better for the number of buckets per partition to be ceiling'ed to minimize the chance that any bucket has more rows than `batchSize`.

Added a test to verify the low-level functionality for calculating the number of partitions and buckets based on row count, the user-defined partition count, and the batch size.