Added splitter test and some messages about embedding customizations
For XML, it doesn't matter if the element already exists since we're doing an `appendChild`.

For JSON - it does matter if the property name already exists. The splitter then uses "splitter-chunks" as the name. I think we can take the same approach for the classifier so we avoid the need for two more options. May later change the embedder to follow the same pattern.