MLE-19467 No need for a subproject for the UDF
The tika-core dependency is so "light", it's simple to include in the connector. Also learned that Tika can be configured via an environment variable, so as of now, we don't need to provide any options for it.

Also using a better way of detecting usage of DocumentRowSchema. This will allow us to freely add more columns via UDFs without breaking anything.