Classifier tests now use a mock
Turned TextClassifier into an interface to support this, with SemaphoreTextClassifier having the goods in it now. Added a few more S4 config tests to ensure we build a connection correctly, as none of the tests now depend on an active S4 connection. A couple tests still have the "live" options for easy manual testing though.

Removed the two disabled UTF-8 tests as we no longer have any code that constructs a String from bytes.