MLE-20029 Can now classify extracted text and chunks together
Still need to add an option for the classifier batch size, but this use case hadn't been addressed yet.

Key changes are in `DocBuilder`, which will now add a classification to the extracted text doc if it exists.

Also swapped out one test PDF for a smaller one for easier testing.