MLE-26379 Added debug logging for deserialization failures
Just a logging addition, no other change in functionality. The logging is intended to be available when a problem like the one in this ticket pops up again, where the deserialization failure actually held an important clue, but the error wasn't logged. Most of the time, the error won't be interesting but will be expected - hence, it's at the debug level.

Also added a small optimization - so no change in functionality either, just less wasteful - where we don't try to deserialize if we know the format is XML.